requirements.txt
arrow==1.2.3
asgiref==3.5.2
dateparser==1.1.1
Django==4.1.2
django-staticfiles-moment==2.0.0
djangorestframework==3.14.0
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
six==1.16.0
sqlparse==0.4.3
times==0.7
tzdata==2022.4
tzlocal==4.2
